requirements.txt
# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Google Cloud Vision API
google-cloud-vision==3.10.2
google-generativeai==0.8.3

# Notion API
notion-client==2.2.1

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow>=10.0.0

# Image Processing
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Logging
structlog==24.1.0

# System monitoring
psutil==5.9.6
